refactor: dsl/ モジュール切り出し - Parser -> Evaluator - Config -> ConfigContext - テストコードもディレクトリ構造を合わせた
