--- src/app/lib/domain_util.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class DomainUtil
-  def self.normalize_name(name)
-    name.gsub(/[\.\W]/, '_')
-  end
-end
