+++ src/app/lib/dsl/domain_util.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Dsl
+  class DomainUtil
+    def self.normalize_name(name)
+      name.gsub(/[\.\W]/, '_')
+    end
+  end
+end
